feat: update quantity field type to Double in ProductUnitOfMeasure DTOs, model, and database schema

+++ backend/mysell/src/main/java/com/project/mysell/dto/product/unit/ProductUnitOfMeasureResponseDTO.java
@@ -3,7 +3,7 @@ package com.project.mysell.dto.product.unit;
 import com.project.mysell.dto.unit.UnityOfMeasureDTO;
 
 public record ProductUnitOfMeasureResponseDTO(
-	    Long quantity,
+	    Double quantity,
 	    UnityOfMeasureDTO unityOfMeasure
 		) {
